backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
asgiref==3.9.2
    # via django
django==5.2.7
    # via
    #   -r requirements.in
    #   django-ninja
django-ninja==1.4.3
    # via -r requirements.in
nodeenv==1.9.1
    # via pyright
pydantic==2.11.9
    # via
    #   -r requirements.in
    #   django-ninja
pydantic-core==2.33.2
    # via pydantic
pyright==1.1.406
    # via -r requirements.in
sqlparse==0.5.3
    # via django
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   pyright
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
